Adds system main block as an ombublock and allows per context widths
Widths can now be set on a per context basis (needed to allow the width to be
set on the global system main block on a per-page basis).  This means that
ombublocks_block_set_width() can no longer be called when creating a new bean,
unless you pass in the context name as well.